공지 등록 중간 저장

+++ src/main/java/com/wisehr/wisehr/notice/dto/NotMemberDTO.java
@@ -8,7 +8,7 @@ import lombok.*;
 @Getter
 @ToString
 public class NotMemberDTO {
-    private int memCode;
+    private Long memCode;
     private String memName;
     private String memPhone;
     private String memEmail;
@@ -19,6 +19,6 @@ public class NotMemberDTO {
     private String memEndDate;
     private String memStatus;
     private String memRoll;
-    private int posCode;
-    private int depCode;
+    private Long posCode;
+    private Long depCode;
 }
